requirements.txt
aiohttp
aiolimiter
beautifulsoup4
PyYAML